move FileExists to internal package
